+++ flutter_inappwebview_forge_platform_interface/lib/src/in_app_localhost_http_request.dart
@@ -1,0 +1,2 @@
+export 'in_app_localhost_http_request_stub.dart'
+    if (dart.library.io) 'in_app_localhost_http_request_io.dart';